% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfication_accuracy_mb.R
\name{classfication_accuracy_mb}
\alias{classfication_accuracy_mb}
\title{A function to calculate the classification accuracy of a Markov blanket learner}
\usage{
classfication_accuracy_mb(mbTrue, mbLearned, vars, target)
}
\arguments{
\item{mbTrue}{The true Markov blanket}

\item{mbLearned}{The learned Markov blanket}

\item{vars}{A vector of all variables.}

\item{target}{The target node.}
}
\description{
This is a function to calculate the classification accuracy of a Markov blanket learner. In details, 
the function calculates TP, TN, FP, FN, and returns precision and recall as outputs. If true mb is
empty but learned mb is not, then both precision and recall are 0; if learned mb is empty but true 
mb is not, then both precision and recall are 0; if both true and learned mbs are empty, then both
precision and recall are 1.
}
