% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag2matrix.R
\name{dag2matrix}
\alias{dag2matrix}
\title{A function to convert a dag from the bnlearn format to an adjacency matrix}
\usage{
dag2matrix(dag)
}
\description{
This converts a dag stored in the bnlearn format to an adjacency matrix contains 0s and 1s. A directed arc 
V1 -> V2 has a corresponding value 1 at the matrix entry [v1, v2]. Matrix column and row names are in the order
of bnlearn::nodes(dag).
}
