% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim_nb.R
\name{fim_nb}
\alias{fim_nb}
\title{Calculate FIM of NB}
\usage{
fim_nb(prodPij1, prodPij0, px, probsMatrix, py1, py0, arities, xIndices, yIndex)
}
\arguments{
\item{prodPij1}{A vector of \prod_j p(x_ij|y=1).}

\item{prodPij0}{A vector of \prod_j p(x_ij|y=0).}

\item{px}{A vector of py1 * prodPij1 + py0 * prodPij0.}

\item{probsMatrix}{A matrix of p(x_j|y=1) and p(y_j|y=0)}

\item{py1}{p(y=T)}

\item{py0}{p(y=F)}

\item{arities}{A vector of variable arities.}

\item{xIndices}{A vector of input variables indices.}

\item{yIndex}{The index of the output variable.}
}
\description{
This function calculates the expected fisher information matrix of a Naive Bayes model.
}
