% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_cpt_fast.R
\name{mml_cpt_fast}
\alias{mml_cpt_fast}
\title{A faster version of mml_cpt}
\usage{
mml_cpt_fast(indexListPerNodePerValue, cachedIndicesList, arities, sampleSize,
  parentsIndices, targetIndex, logFactorialSheet, base)
}
\arguments{
\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained 
by the function count_occurance().}

\item{cachedIndicesList}{A vector of indices stored to speed up calculations.}

\item{arities}{A vector of variable ariteis.}

\item{sampleSize}{Sample size of a given data set.}

\item{parentsIndices}{Indices of parents nodes.}

\item{targetIndex}{Index of the target node.}

\item{logFactorialSheet}{A pre-saved log factorial sheet for non-negative integers <= 10000.}

\item{base}{The base of the logarithm.}
}
\description{
There is no difference between mml_cpt_fast() and mml_cpt(), other than the speed difference.
}
