% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_fixed_str_adaptive.R
\name{mml_fixed_str_adaptive}
\alias{mml_fixed_str_adaptive}
\title{MML for a fixed Markov blanket model using adaptive code approach}
\usage{
mml_fixed_str_adaptive(data, vars, arities, sampleSize, targetIndex,
  targetProbsAdpt, str)
}
\arguments{
\item{data}{A categorical data set.}

\item{vars}{A vector of all variables in data, in the same order as the column names of data.}

\item{arities}{A vector of variable arities in data, in the same order as the column names of data.}

\item{sampleSize}{The sample size. That is, the number of rows of data.}

\item{targetIndex}{The target node's index in vars, whose Markov blanket we are interested in.}

\item{targetProbsAdpt}{A matrix that stores the target's probability for each of its value at each 
data point. The matrix has dimension arity(target) by sampleSize.}

\item{str}{A fixed Markov blanket model stores in the bnlearn dag format.}
}
\value{
The function outputs the message length of a fixed structure.
}
\description{
This function calculates the mml score of the target under a fixed Markov blanket model. The structure 
can be a general DAG. This returns the 2nd part of MML that is the likelihood of data given the 
structure, since the MML cost of the structure is not stated.
}
\keyword{This}
\keyword{cond_probs_adaptive(),}
\keyword{dependencies}
\keyword{function}
\keyword{has}
\keyword{log_prob_adaptive().}
\keyword{on}
