% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_with_parents_fast.R
\name{mml_with_parents_fast}
\alias{mml_with_parents_fast}
\title{Auxiliary function to mml_cpt_fast}
\usage{
mml_with_parents_fast(indexListPerNodePerValue, cachedIndicesList, arities,
  sampleSize, parentsIndices, targetIndex)
}
\arguments{
\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained
by the function count_occurance().}

\item{cachedIndicesList}{A vector of indices stored to speed up calculations.}

\item{arities}{A vector of varaible arities.}

\item{sampleSize}{Sample size of a given data set.}

\item{parentsIndices}{Indices of parents nodes.}

\item{targetIndex}{Index of the target node.}
}
\description{
This is an auxiliary function to mml_cpt_fast. It calculates the mml score of a node given its parents.
}
