% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_without_parents.R
\name{mml_without_parents}
\alias{mml_without_parents}
\title{Auxiliary function to both mml_cpt and mml_cpt_fast}
\usage{
mml_without_parents(indexListPerNodePerValue, arities, sampleSize, targetIndex)
}
\arguments{
\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained
by the function count_occurance().}

\item{arities}{A vector of variable ariteis.}

\item{sampleSize}{Sample size of a given data set.}

\item{targetIndex}{Index of the target node.}
}
\description{
This is an auxiliary function to both mml_cpt() and mml_cpt_fast(). It calculates the mml score of
a node without any parents.
}
