% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_par_cal.R
\name{single_par_cal}
\alias{single_par_cal}
\title{Auxiliary function to mml_with_parents() and mml_with_parents_fast()}
\usage{
single_par_cal(indexListPerNodePerValue, commonParentsIndices, arityChild,
  targetIndex, base)
}
\arguments{
\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained 
by the function count_occurance().}

\item{commonParentsIndices}{}

\item{arityChild}{}

\item{targetIndex}{Index of the target node.}

\item{base}{The base of the logarithm.}
}
\description{
This is an auxiliary function to both mml_with_parents() and mml_with_parents_fast() when the target 
node has only one parent.
}
