% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_greedy_fast.R
\name{forward_greedy_fast}
\alias{forward_greedy_fast}
\title{A faster version of a forward greedy search}
\usage{
forward_greedy_fast(data, indexListPerNodePerValue, arities, sampleSize, node,
  logFactorialSheet, base = exp(1), debug = FALSE)
}
\arguments{
\item{data}{A categorical data set.}

\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained 
by the function count_occurance().}

\item{arities}{A vector of arities.}

\item{sampleSize}{Sample size of the given data.}

\item{node}{The target node.}

\item{logFactorialSheet}{A pre-saved log factorial sheet for non-negative integers <= 10000.}

\item{base}{The base of the logarithm. Default is the natural log.}

\item{debug}{A boolearn argument to display more details.}
}
\description{
This is a faster version of a forward greedy search. It can only be used with mml_cpt_fast() 
as the objective function.
}
