% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_with_parents.R
\name{mml_with_parents}
\alias{mml_with_parents}
\title{Auxiliary function to mml_cpt()}
\usage{
mml_with_parents(indexListPerNodePerValue, arities, sampleSize, parentsIndices,
  targetIndex, base)
}
\arguments{
\item{indexListPerNodePerValue}{}

\item{arities}{arities}

\item{sampleSize}{sampleSize}

\item{parentsIndices}{parentsIndices}

\item{targetIndex}{targetIndex}

\item{base}{base}
}
\description{
This function calculates the mml score of a target node given its parents. There has to be at least one 
parent for the target.
}
